+++ source/nvp-io/image_io.h
@@ -1,0 +1,15 @@
+/*
+ * This file is part of nodevmap open source project.
+ *
+ * For conditions of distribution and use, see copyright notice in LICENSE file.
+ *
+ * For more detailed information : https://github.com/moveplaten
+ */
+
+#pragma once
+
+class NvpImageIO
+{
+public:
+    static void outputAll(const char* file_name);
+};