+++ source/port-win/CHANGELOG.md
@@ -6,3 +6,6 @@ nodevmap
 
 `2021-10-15`
 * DirectWrite is first choice to support basic text rendering
+
+`2022-01-19`
+* add bitmap render target for final image format output